feat(google): implement external accounts (#308)
Hi! Google introduced a new way of retrieving access tokens and signing
payloads in mid-2022 called [External
Accounts](https://google.aip.dev/auth/4117).
An external account is used by the Workload Identity Federation to
temporarily gain access to a service account
without having to distribute credential JSONs everywhere. This is done
by having the auth be handled between
the OIDC (AWS, Azure, Github, etc) provider and Google directly.

That said, the scheme is also vastly different from how service account
auths work, but will probably see a rise
in adoption in the coming months. For example, a user using Sccache on
their Github Actions might want to log
in to GCS using a token provided by Github (this is my case and what got
me here).

This PR implements this new scheme, supporting URL and file-sourced
tokens, but still missing implementations
for the other two available sources: environmental (used by AWS
environments) and executable sources. These
two can come in a future PR due to their complexity and necessity for
attention.

## Summary

* Creates a struct called `ExternalAccount`;
* Implements two credential sources: file and URL sources;
* Creates an enum called `CredentialAccount` with two variants:
`ExternalAccount` and `ServiceAccount`;
* Adds a `fn load_account(): CredentialAccount` to `CredentialLoader`;
* Adds a `fn with_account(CredentialAccount)` to `TokenLoader`;
* Renames `Credential` to `ServiceAccount` but leaves `GoogleCredential`
as an export alias;
* Both `CredentialLoader` and `TokenLoader` use `CredentialAccount`
instead of `Credential` for their underlying type.

## Unresolved questions

### 1. What should the interface look like?
There's currently a single struct called `Credential` which only
represents a service account JSON, but it may be
misleading as it isn't generic, as the name implies.

To avoid making a huge breaking change, I'd suggest creating an
`Account` enum holding service or external
accounts, and add `#[deprecated]` for items that depend on `Credential`,
and rename `Account` to `Credential`
on a future release.

My preferred solution would be to make the `Credential` struct be said
enum, but this would be a huge breaking
change for the users.

### 2. How should the function `Signer::sign_query` be implemented?
It isn't possible to locally sign a blob via external accounts as they
don't hold the account's private key. It's possible
however to sign by POSTing to
[`projects.serviceAccounts.signBlob`](https://cloud.google.com/iam/docs/reference/credentials/rest/v1/projects.serviceAccounts/signBlob).
Some solutions include:
* `sign_query` could be turned into an async function, but this is
undesirable so as to not make things complicated;
* `sign_query` could only support service account credentials, but I
consider this gatekeeping;
* `sign_query` could remain being sync and use `reqwest::blocking` when
an external account is used;
* Together with the solution above, there could be a counterpart
`sign_query_async` using non-blocking `reqwest`.

### 3. How should it be tested?
Testing this solution requires having a valid OIDC token, the kind that
providers calculate, like Github.
The tests would only work on the Github CI, and creating a Workload
Identity Federation Pool would be
necessary. [More on
that](https://medium.com/google-cloud/how-does-the-gcp-workload-identity-federation-work-with-github-provider-a9397efd7158).