🦙 Azure Storage Add AAD OAuth support (#322)
This PR is to add AAD OAuth support for Azure Storage. Since AAD OAuth
essentially authorizes through Access Tokens, we directly added a
BearerToken in the Credential.

Please note that we have not implemented how to obtain the Access Token,
and this part needs to be implemented by the user themselves.

close #164  ?

maybe we can use https://github.com/Azure/login action to test this
changes in the feature.